refactor login screen
